schedule/requirements.txt
django-nose>=1.3
nose>=1.3.4
coveralls>=0.4.4
